% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_full_dataset.R
\name{download_full_dataset}
\alias{download_full_dataset}
\title{importFrom downloader dwonload}
\usage{
download_full_dataset(sel_dataset, format = "shp")
}
\description{
importFrom downloader dwonload
}
