% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Retrieves the actual data from a selected dataset.
TODO: Where clauses.}
\usage{
get_dataset(sel_dataset, fields = FALSE, sp_ref = 4326,
  export = FALSE, format = "shp")
}
\arguments{
\item{format}{}
}
\description{
Retrieves the actual data from a selected dataset.
TODO: Where clauses.
}
\keyword{get_dataset}
