% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dataset.R
\name{select_dataset}
\alias{select_dataset}
\title{Selects a dataset from a search_dataset response.}
\usage{
select_dataset(dataset_df, id)
}
\arguments{
\item{id}{}
}
\description{
Selects a dataset from a search_dataset response.
}
\keyword{select_dataset}
