% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_aoi}
\alias{set_aoi}
\title{Sets an aoi for the current session.}
\usage{
set_aoi(sp_obj = FALSE)
}
\arguments{
\item{sp_obj}{}
}
\description{
Sets an aoi for the current session.
}
\keyword{settings}
