% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{fields}
\alias{fields}
\title{Returns the field names of a selected dataset.}
\usage{
fields(sel_dataset)
}
\arguments{
\item{sel_dataset}{}
}
\description{
Returns the field names of a selected dataset.
}
\keyword{fields}
