% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datasets.R
\name{search_datasets}
\alias{search_datasets}
\title{Search datasets}
\usage{
search_datasets(category = c(), name = "", source = "",
  sector = "", region = "", aoi = FALSE)
}
\description{
Search datasets
}
